Optimize solution to account for stack overflow

+++ src/app/Factorial.java
@@ -1,13 +1,13 @@
 package app;
 
 public class Factorial {
-    int result;
+    long result;
     
     public Factorial(int n) {
         result = calculate(n);
     }
 
-    private int calculate(int n) {
+    private long calculate(int n) {
         if (n == 1)
             return 1;
